fix(bfi): use byte cells
All the other examples work with 8-bit or 32-bit cells, but LostKng.b
requires 8-bit cells. This change brings the interpreter in line with
the accepted 8-bit convention.

+++ bfi.h
@@ -13,6 +13,6 @@
 #include "parser.h"
 #include "passes.h"
 
-int32_t bf_interpret(bf_ast *ast, int32_t *memory, int32_t dptr);
+intptr_t bf_interpret(bf_ast *ast, uint8_t *memory, intptr_t dptr);
 
 #endif
